refactor(relax): extract socket frame math helpers into socket_frame_utils
Move Jacobi SVD and matrix utility functions from socket_frame.cpp's
anonymous namespace into a dedicated socket_frame_utils module. This
reduces socket_frame.cpp from 426 to 278 lines, making the remaining
validation logic easier to decompose in follow-up changes.

Update CMakeLists.txt and test/CMakeLists.txt to include the new source
file in the library and both socket test targets.
